+++ src/day5.rs
@@ -2,24 +2,24 @@ use crate::machine::Machine;
 use aoc_runner_derive::{aoc, aoc_generator};
 
 #[aoc_generator(day5)]
-fn parse_input(input: &str) -> Vec<i32> {
+fn parse_input(input: &str) -> Vec<i64> {
     input
         .split(|c| c == ',' || c == '\r')
         .filter(|c| c != &"")
-        .map(|line| line.parse::<i32>().unwrap())
+        .map(|line| line.parse::<i64>().unwrap())
         .collect()
 }
 #[aoc(day5, part1)]
-fn part1(instructions: &[i32]) -> i32 {
+fn part1(instructions: &[i64]) -> i64 {
     run(instructions, &[1])
 }
 
 #[aoc(day5, part2)]
-fn part2(instructions: &[i32]) -> i32 {
+fn part2(instructions: &[i64]) -> i64 {
     run(instructions, &[5])
 }
 
-fn run(instructions: &[i32], inputs: &[i32]) -> i32 {
+fn run(instructions: &[i64], inputs: &[i64]) -> i64 {
     let mut m = Machine::new(instructions.to_vec(), inputs.to_vec());
     m.process();
     for o in m.outputs() {
